+++ static/libs/markdown.js
@@ -1,1 +1,1 @@
-export{w as init}from"./markdown-B2YDq4LV.js";
+export{w as init}from"./markdown-B8R4CVWq.js";